# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Budget Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /income:
    get:
      operationId: listIncome
      tags: [income]
      summary: List income entries
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "Format: YYYY-MM"
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of income entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncomeEntry"
    post:
      operationId: createIncome
      tags: [income]
      summary: Create income entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeBody"
      responses:
        "201":
          description: Created income entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
  /income/{id}:
    put:
      operationId: updateIncome
      tags: [income]
      summary: Update income entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeBody"
      responses:
        "200":
          description: Updated income entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
    delete:
      operationId: deleteIncome
      tags: [income]
      summary: Delete income entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expense entries
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "Format: YYYY-MM"
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of expense entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseEntry"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseEntry"
  /expenses/{id}:
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "200":
          description: Updated expense entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseEntry"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /savings-goals:
    get:
      operationId: listSavingsGoals
      tags: [savings-goals]
      summary: List savings goals
      responses:
        "200":
          description: List of savings goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavingsGoal"
    post:
      operationId: createSavingsGoal
      tags: [savings-goals]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavingsGoalBody"
      responses:
        "201":
          description: Created savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
  /savings-goals/{id}:
    put:
      operationId: updateSavingsGoal
      tags: [savings-goals]
      summary: Update savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavingsGoalBody"
      responses:
        "200":
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
    delete:
      operationId: deleteSavingsGoal
      tags: [savings-goals]
      summary: Delete savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /savings-goals/{id}/contribution:
    post:
      operationId: addSavingsContribution
      tags: [savings-goals]
      summary: Add a contribution to a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddContributionBody"
      responses:
        "200":
          description: Updated savings goal after contribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
...
components:
  schemas:
...
    IncomeEntry:
      type: object
      properties:
        id:
          type: integer
        source:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        categoryId:
          type: integer
        categoryName:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - source
        - amount
        - date
        - categoryId
        - categoryName
        - createdAt
    CreateIncomeBody:
      type: object
      properties:
        source:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        categoryId:
          type: integer
        notes:
          type: string
      required:
        - source
        - amount
        - date
        - categoryId
    ExpenseEntry:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryIcon:
          type: string
        categoryColor:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - description
        - amount
        - date
        - categoryId
        - categoryName
        - categoryIcon
        - categoryColor
        - createdAt
    CreateExpenseBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        categoryId:
          type: integer
        notes:
          type: string
      required:
        - description
        - amount
        - date
        - categoryId
...
    SavingsGoal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
          format: date
        notes:
          type: string
        percentComplete:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - targetAmount
        - currentAmount
        - percentComplete
        - createdAt
    CreateSavingsGoalBody:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
          format: date
        notes:
          type: string
      required:
        - name
        - targetAmount
        - currentAmount
    AddContributionBody:
      type: object
      properties:
        amount:
          type: number
        notes:
          type: string
      required:
        - amount
...