# lib/api-zod/src/generated/types/createExpenseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseBody {
  description: string;
  amount: number;
  date: Date;
  categoryId: number;
  notes?: string;
}