# lib/api-zod/src/generated/types/createIncomeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateIncomeBody {
  source: string;
  amount: number;
  date: Date;
  categoryId: number;
  notes?: string;
}