# lib/api-zod/src/generated/types/expenseEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseEntry {
  id: number;
  description: string;
  amount: number;
  date: Date;
  categoryId: number;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  notes?: string;
  createdAt: Date;
}