# lib/api-zod/src/generated/types/incomeEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface IncomeEntry {
  id: number;
  source: string;
  amount: number;
  date: Date;
  categoryId: number;
  categoryName: string;
  notes?: string;
  createdAt: Date;
}